refactor: _safe_l2_normalizeの重複実装をVectorUtilsクラスに統合
image_quality_analyzer.pyとfeature_extractor.pyで重複していた_safe_l2_normalize関数を、
src/utils/vector_utils.pyのVectorUtilsクラスとして統合しました。
テストもtests/test_vector_utils.pyに一本化し、重複を解消しました。
