+++ src/utils/__init__.py
@@ -1,5 +1,6 @@
-"""Utils package for file operations."""
+"""Utils package for file operations and vector utilities."""
 
 from .file_utils import FileUtils
+from .vector_utils import VectorUtils
 
-__all__ = ["FileUtils"]
+__all__ = ["FileUtils", "VectorUtils"]
